+++ app/sprinkles/admin/assets/userfrosting/js/pages/permission.js
@@ -12,4 +12,9 @@ $(document).ready(function() {
         dataUrl: site.uri.public + '/api/permissions/p/' + page.permission_id + '/users',
         useLoadingTransition: site.uf_table.use_loading_transition
     });
+
+    // Bind table buttons
+    $("#widget-permission-users").on("pagerComplete.ufTable", function () {
+        bindUserButtons($(this));
+    });
 });